+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+
+### Fixed
+- Fix bug in error handling where recoverable errors made the integration to panic.
+
 ### 1.0.0-beta2.1
 
 ### Added